Moving option to separate file
