+++ Xsd2/Xsd2.csproj
@@ -52,6 +52,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="XsdCodeGenerator.cs" />
+    <Compile Include="XsdCodeGeneratorOptions.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\NuGet\Xsd2\Codaxy.Xsd2.nuspec">
