+++ Xsd2/XsdCodeGenerator.cs
@@ -544,28 +544,4 @@ namespace Xsd2
             return p.Substring(0, 1).ToLower() + p.Substring(1) + suffix;
         }
     }
-
-    public class XsdCodeGeneratorOptions
-    {
-        public bool UseLists { get; set; }
-        public bool CapitalizeProperties { get; set; }
-        public bool CapitalizeEnumValues { get; set; }
-        public bool StripDebuggerStepThroughAttribute { get; set; }
-        public bool StripPclIncompatibleAttributes { get; set; }
-        public bool HideUnderlyingNullableProperties { get; set; }
-        public bool UseNullableTypes { get; set; }
-        public List<String> Imports { get; set; }
-        public List<String> UsingNamespaces { get; set; }
-        
-        public string OutputNamespace { get; set; }
-        public bool WriteFileHeader { get; set; }
-        public bool PreserveOrder { get; set; }
-        public bool EnableDataBinding { get; set; }
-
-        public bool ExcludeImportedTypes { get; set; }
-
-        public bool ExcludeImportedTypesByNameAndNamespace { get; set; }
-
-        public bool MixedContent { get; set; }
-    }
 }
